+++ Common/Api/LiveAlgorithmResultsJsonConverter.cs
@@ -114,7 +114,7 @@ namespace QuantConnect.Api
 
             // Live Results - At this time only that charting data can be returned from the api (9/30/2016)
             liveAlgoResults.LiveResults.Results = new LiveResult(new LiveResultParameters(chartDictionary,
-                new Dictionary<int, Order>(),
+                new Dictionary<long, Order>(),
                 new Dictionary<DateTime, decimal>(),
                 new Dictionary<string, Holding>(),
                 new CashBook(),
