+++ Common/Interfaces/IAlgorithm.cs
@@ -720,7 +720,7 @@ namespace QuantConnect.Interfaces
         /// <param name="symbolToLiquidate">Specific asset to liquidate, defaults to all.</param>
         /// <param name="tag">Custom tag to know who is calling this.</param>
         /// <returns>list of order ids</returns>
-        List<int> Liquidate(Symbol symbolToLiquidate = null, string tag = "Liquidated");
+        List<long> Liquidate(Symbol symbolToLiquidate = null, string tag = "Liquidated");
 
         /// <summary>
         /// Set live mode state of the algorithm run: Public setter for the algorithm property LiveMode.
