+++ Common/Packets/LiveResultPacket.cs
@@ -135,7 +135,7 @@ namespace QuantConnect.Packets
         public static LiveResultPacket CreateEmpty(LiveNodePacket job)
         {
             return new LiveResultPacket(job, new LiveResult(new LiveResultParameters(
-                new Dictionary<string, Chart>(), new Dictionary<int, Order>(), new Dictionary<DateTime, decimal>(),
+                new Dictionary<string, Chart>(), new Dictionary<long, Order>(), new Dictionary<DateTime, decimal>(),
                 new Dictionary<string, Holding>(), new CashBook(), new Dictionary<string, string>(),
                 new SortedDictionary<string, string>(), new List<OrderEvent>(), new Dictionary<string, string>(),
                 new AlgorithmConfiguration(), new Dictionary<string, string>())));
